Add very simple raylib example
Currently everything to do with raylib is done in the native exec
with only the request to start the gui done on the server.
The next step is to seperately wrap all of the required raylib
functions and pass the required data through the pipes. Eventually
raylib will be entirely in the cosmo exec but small steps.