Get raylib working for windows and linux with a single binary
Currently raylib is in a seperate shared libary for each platform but I intend
to move it into the main exec with only glfw in a shared library instead.
Most of the work to make this happen is working around cosmo's limited parameter
support for dynamically loaded functions with 5 integer parameters and recently
6 floats. I do have a modified version that supports 6 integers params but have
not solved the stack issue required to support more scalar params and passing
structs by value like raylib does.