Partially resolve patch issues in pull
On pull the patches can obviously change but the applied patches
do not which causes issues rebuilding as the patch reapplication
fails. A better fix would be to not reapply the patches if there
are any local changes but for now running `make distclean` after
any pulls to reset the submodules does work.