Improve parallel nano-X build prevention
Since nano-X builds in tree the windows and linux builds cannot
occur at the same time. Previously I had made the windows build
dependent on the linux build but that resulted in both building
during a windows build. Now the build script checks if both are
requested and only then makes one dependent on the other.